feat: userId null 예외처리 구현

+++ src/backend/state-service/stateservice/src/main/java/harmony/stateservice/service/ChatServerService.java
@@ -1,6 +1,7 @@
 package harmony.stateservice.service;
 
 import harmony.stateservice.dto.SessionDto;
+import harmony.stateservice.dto.response.BaseResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.HashOperations;
@@ -14,26 +15,33 @@ import org.springframework.stereotype.Service;
 public class ChatServerService {
 
     private final RedisTemplate<String, Object> redisTemplate;
+    private final ResponseService responseService;
 
-    public String updateSession(SessionDto sessionDto) {
+    public BaseResponse<String> updateSession(SessionDto sessionDto) {
         HashOperations<String, Object, Object> hashOperations = redisTemplate.opsForHash();
         String sessionId = sessionDto.getSessionId();
         String stateKey = "USER:STATE";
         String userId = null;
         if (sessionDto.getType().equals("CONNECT")) {
             userId = String.valueOf(sessionDto.getUserId());
+            if (userId == null) {
+                return responseService.getFailResponse(null);
+            }
             hashOperations.put(stateKey, userId, sessionDto.getState());
             saveSessionId(sessionId, userId);
         }
         if (sessionDto.getType().equals("DISCONNECT")) {
             userId = getUserId(sessionId);
+            if (userId == null) {
+                return responseService.getFailResponse(null);
+            }
             hashOperations.put(stateKey, userId, sessionDto.getState());
             deleteSessionId(sessionId);
         }
 
         log.info("userId {}", userId);
-        log.info("user state {}", hashOperations.get(stateKey, userId));
-        return userId;
+        log.info("connection {}", sessionDto.getState());
+        return responseService.getSuccessResponse(userId);
     }
 
     public void saveSessionId(String sessionId, String userId) {
@@ -47,6 +55,11 @@ public class ChatServerService {
 
     public String getUserId(String sessionId) {
         ValueOperations<String, Object> valueOperations = redisTemplate.opsForValue();
-        return valueOperations.get(sessionId).toString();
+        Object checkUserId = valueOperations.get(sessionId);
+        if (checkUserId != null) {
+            return checkUserId.toString();
+        }
+
+        return null;
     }
 }
